# storage: decouple the region storage from LevelDB backend