--- a/pkg/storage/leveldb_backend.go
+++ b/pkg/storage/leveldb_backend.go
@@ -38,19 +38,13 @@ const (
 	defaultDirtyFlushTick = time.Second
 )
 
-type levelDBKV interface {
-	Key() string
-	Value() ([]byte, error)
-	Encrypt(*encryption.Manager) (levelDBKV, error)
-}
-
 // levelDBBackend is a storage backend that stores data in LevelDB,
 // which is mainly used to store the PD Region meta information.
-type levelDBBackend[T levelDBKV] struct {
+type levelDBBackend struct {
 	*endpoint.StorageEndpoint
 	ekm       *encryption.Manager
 	mu        syncutil.RWMutex
-	batch     map[string]T
+	batch     map[string][]byte
 	batchSize int
 	cacheSize int
 	flushRate time.Duration
@@ -60,21 +54,21 @@ type levelDBBackend[T levelDBKV] struct {
 }
 
 // newLevelDBBackend is used to create a new LevelDB backend.
-func newLevelDBBackend[T levelDBKV](
+func newLevelDBBackend(
 	ctx context.Context,
 	filePath string,
 	ekm *encryption.Manager,
-) (*levelDBBackend[T], error) {
+) (*levelDBBackend, error) {
 	levelDB, err := kv.NewLevelDBKV(filePath)
 	if err != nil {
 		return nil, err
 	}
-	lb := &levelDBBackend[T]{
+	lb := &levelDBBackend{
 		StorageEndpoint: endpoint.NewStorageEndpoint(levelDB, ekm),
 		ekm:             ekm,
 		batchSize:       defaultBatchSize,
 		flushRate:       defaultFlushRate,
-		batch:           make(map[string]T, defaultBatchSize),
+		batch:           make(map[string][]byte, defaultBatchSize),
 		flushTime:       time.Now().Add(defaultFlushRate),
 	}
 	lb.ctx, lb.cancel = context.WithCancel(ctx)
@@ -82,7 +76,7 @@ func newLevelDBBackend[T levelDBKV](
 	return lb, nil
 }
 
-func (lb *levelDBBackend[T]) backgroundFlush() {
+func (lb *levelDBBackend) backgroundFlush() {
 	defer logutil.LogPanic()
 
 	var (
@@ -112,72 +106,55 @@ func (lb *levelDBBackend[T]) backgroundFlush() {
 	}
 }
 
-func (lb *levelDBBackend[T]) SaveKV(kv T) error {
-	key := kv.Key()
-	value, err := kv.Encrypt(lb.ekm)
-	if err != nil {
-		return err
-	}
+// SaveInBatch saves the key-value pair in the batch cache, and it will
+// only be saved to the underlying storage when the `Flush` method is
+// called or the cache is full.
+func (lb *levelDBBackend) SaveInBatch(key string, value []byte) error {
 	lb.mu.Lock()
 	defer lb.mu.Unlock()
 	if lb.cacheSize < lb.batchSize-1 {
-		lb.batch[key] = value.(T)
+		lb.batch[key] = value
 		lb.cacheSize++
 
 		lb.flushTime = time.Now().Add(lb.flushRate)
 		return nil
 	}
-	lb.batch[key] = value.(T)
-	err = lb.flushLocked()
-
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (lb *levelDBBackend[T]) DeleteKV(kv T) error {
-	return lb.Remove(kv.Key())
+	lb.batch[key] = value
+	return lb.flushLocked()
 }
 
 // Flush saves the batch cache to the underlying storage.
-func (lb *levelDBBackend[T]) Flush() error {
+func (lb *levelDBBackend) Flush() error {
 	lb.mu.Lock()
 	defer lb.mu.Unlock()
 	return lb.flushLocked()
 }
 
-func (lb *levelDBBackend[T]) flushLocked() error {
+func (lb *levelDBBackend) flushLocked() error {
 	if err := lb.saveBatchLocked(); err != nil {
 		return err
 	}
 	lb.cacheSize = 0
-	lb.batch = make(map[string]T, lb.batchSize)
+	lb.batch = make(map[string][]byte, lb.batchSize)
 	return nil
 }
 
-func (lb *levelDBBackend[T]) saveBatchLocked() error {
+func (lb *levelDBBackend) saveBatchLocked() error {
 	batch := new(leveldb.Batch)
-
-	for k, v := range lb.batch {
-		value, err := v.Value()
-		if err != nil {
-			return err
-		}
-		batch.Put([]byte(k), value)
+	for key, value := range lb.batch {
+		batch.Put([]byte(key), value)
 	}
-
 	if err := lb.Base.(*kv.LevelDBKV).Write(batch, nil); err != nil {
 		return errs.ErrLevelDBWrite.Wrap(err).GenWithStackByCause()
 	}
 	return nil
 }
 
-// Close closes the LevelDB kv. It will call Flush() once before closing.
-func (lb *levelDBBackend[T]) Close() error {
+// Close will gracefully close the LevelDB backend and flush the data to the underlying storage before closing.
+func (lb *levelDBBackend) Close() error {
 	err := lb.Flush()
 	if err != nil {
-		log.Error("meet error before close the leveldb storage", errs.ZapError(err))
+		log.Error("meet error before closing the leveldb storage", errs.ZapError(err))
 	}
 	lb.cancel()
 	err = lb.Base.(*kv.LevelDBKV).Close()