--- a/pkg/storage/storage.go
+++ b/pkg/storage/storage.go
@@ -57,13 +57,18 @@ func NewStorageWithEtcdBackend(client *clientv3.Client, rootPath string) Storage
 	return newEtcdBackend(client, rootPath)
 }
 
-// NewStorageWithLevelDBBackend creates a new storage with LevelDB backend.
-func NewStorageWithLevelDBBackend(
+// NewRegionStorageWithLevelDBBackend will create a specialized storage to
+// store region meta information based on a LevelDB backend.
+func NewRegionStorageWithLevelDBBackend(
 	ctx context.Context,
 	filePath string,
 	ekm *encryption.Manager,
-) (Storage, error) {
-	return newLevelDBBackend(ctx, filePath, ekm)
+) (*RegionStorage, error) {
+	levelDBBackend, err := newLevelDBBackend[*regionKV](ctx, filePath, ekm)
+	if err != nil {
+		return nil, err
+	}
+	return newRegionStorage(levelDBBackend), nil
 }
 
 // TODO: support other KV storage backends like BadgerDB in the future.
@@ -88,12 +93,13 @@ func NewCoreStorage(defaultStorage Storage, regionStorage endpoint.RegionStorage
 	}
 }
 
-// TryGetLocalRegionStorage gets the local region storage. Returns nil if not present.
+// TryGetLocalRegionStorage gets the local region storage and returns nil if not present.
+// It's used to retrieve the real region storage implementation from the given `Storage“.
 func TryGetLocalRegionStorage(s Storage) endpoint.RegionStorage {
 	switch ps := s.(type) {
 	case *coreStorage:
 		return ps.regionStorage
-	case *levelDBBackend, *memoryStorage:
+	case *levelDBBackend[*regionKV], *memoryStorage:
 		return ps
 	default:
 		return nil