… pkg/storage/region_storage_test.go
… pkg/storage/storage_test.go
… pkg/syncer/client_test.go
… tests/server/region_syncer/region_syncer_test.go