<!--

Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: ref #3453.

### What is changed and how does it work?

<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
In the previous implementation, LevelDB and `RegionStorage` were strongly coupled.
This PR decouples region storage from it to further optimize code abstraction, so we can
get a more reusable `LevelDBBackend` and a more specialized `RegionStorage` implementation.
```

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test
- Integration test
- Manual test (add detailed scripts or steps below)

Benchmark results show no performance regression after this PR.

Before this PR:

```shell
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkLoadRegionsByVolume$ github.com/tikv/pd/pkg/storage

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/storage
BenchmarkLoadRegionsByVolume/input_size_10000-8         	1000000000	         0.01101 ns/op	       0 B/op	       0 allocs/op
BenchmarkLoadRegionsByVolume/input_size_100000-8        	1000000000	         0.1247 ns/op	       0 B/op	       0 allocs/op
BenchmarkLoadRegionsByVolume/input_size_1000000-8       	       1	1679503541 ns/op	956784504 B/op	15258023 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/storage	7.184s

Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkLoadRegionsByRandomMerge$ github.com/tikv/pd/pkg/storage

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/storage
BenchmarkLoadRegionsByRandomMerge/merge_ratio_0-8         	       1	1854904083 ns/op	956702808 B/op	15257770 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_20-8        	       1	2366610458 ns/op	983400432 B/op	17280694 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_40-8        	       1	3095372333 ns/op	1054758448 B/op	19185120 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_60-8        	       1	3557470000 ns/op	1086592696 B/op	21083199 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_80-8        	       1	4101661666 ns/op	1135966808 B/op	22915303 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/storage	26.208s
```

After this PR:

```shell
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkLoadRegionsByVolume$ github.com/tikv/pd/pkg/storage

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/storage
BenchmarkLoadRegionsByVolume/input_size_10000-8         	1000000000	         0.01097 ns/op	       0 B/op	       0 allocs/op
BenchmarkLoadRegionsByVolume/input_size_100000-8        	1000000000	         0.1205 ns/op	       0 B/op	       0 allocs/op
BenchmarkLoadRegionsByVolume/input_size_1000000-8       	       1	1678767666 ns/op	956712920 B/op	15257639 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/storage	7.489s

Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkLoadRegionsByRandomMerge$ github.com/tikv/pd/pkg/storage

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/storage
BenchmarkLoadRegionsByRandomMerge/merge_ratio_0-8         	       1	1604244292 ns/op	956722936 B/op	15257692 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_20-8        	       1	2388323625 ns/op	981547472 B/op	17280198 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_40-8        	       1	3082015125 ns/op	1055231792 B/op	19185170 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_60-8        	       1	3575929417 ns/op	1088751552 B/op	21092884 allocs/op
BenchmarkLoadRegionsByRandomMerge/merge_ratio_80-8        	       1	4092867875 ns/op	1135758544 B/op	22912176 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/storage	25.085s
```

### Release note

<!--

A bugfix or a new feature needs a release note. If there is no need to give a release note, just leave it with the `None`.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

-->

```release-note
None.
```